+++ web/src/types/NetParams.d.ts
@@ -199,7 +199,7 @@ declare module NetParams {
 
   interface PostGetPostComments {
     id: number;
-    sort_strategy: "default" | "newest";
+    style: "default" | "hots" | "newest";
     page?: number;
     page_size?: number;
   }